effected comment changes
